+++ common/bolt11.h
@@ -64,6 +64,9 @@ struct bolt11 {
 	/* signature of sha256 of entire thing. */
 	secp256k1_ecdsa_signature sig;
 
+	/* Features bitmap, if any. */
+	u8 *features;
+
 	struct list_head extra_fields;
 };